refactor(cmd): split main.go into cohesive files and add unit tests
Track D. cmd/chatgpt/main.go was 1444 LOC of mixed concerns with no direct
unit tests (only end-to-end integration coverage).

Refactor (behavior-preserving; verified by the existing integration suite that
builds the real binary against a mock server):
- Split package main into cohesive files: main.go (entry + globals), run.go
  (dispatch + interactive/terminal helpers), config.go (viper load/persist/sync),
  agent.go (agent runners), flags.go (cobra flag/help setup). main.go is now
  ~146 LOC.
- Extract 14 pure, global-free helpers into cmd/chatgpt/utils (the existing
  well-tested helper package): BoolToOnOff, ToAliasFlagName, ResolveAgentMode,
  BuildAgentGoal, ParseToolKinds, MergeMaps, FileExists, the flag classifiers,
  and the YAML config read/update/write trio.

Tests (all Unit scope — no binary build, no network; the 40 new cases run in
<1s and the full unit suite stays ~2.5s):
- Table-driven tests for every extracted helper (edge cases: mode aliases,
  tool-kind dedup/errors, YAML comment preservation + scalar-type round-trip,
  merge aliasing, flag classification).
- agent/tools had ZERO tests; add unit tests for FSIOFileOps (write/read,
  ReplaceBytesInFile all/limited/errors, PatchFile) using real fsio against
  t.TempDir, and for ExecShellRunner (stdout/exit-code capture, workdir,
  fast-kill on context cancel) using real harmless commands.
- Add a ClientLLM.Complete test asserting config (OmitHistory/Temperature) is
  restored even when the query errors, and that history is left untouched.

Reviewed with Grok + Gemini: added the llm.go restore-on-error test, documented
UpdateConfigNode's scalar-stringification behavior (+ an int/bool round-trip
test), made the shell-cancellation test assert an early kill, and added
merge-aliasing / whitespace-mode / directory cases.
